+++ FmodPlatformSystem/Assets/Scripts/Scene/SceneMover.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class SceneMover : MonoBehaviour
+{
+    public SceneReference NextScene;
+
+    public void MoveScene()
+    {
+        SceneManager.LoadScene(NextScene.Path);
+    }
+}